chore(1-3336): decouple from usetrafficdata hook
